[feat] 황유빈 / 메인.js 수정 헤더 컴포넌트 스크립트 삭제
